fix unsupported image formats (WebP+avif) by using sharp to process what jimp cannot
