+++ packages/lib/index.ts
@@ -2,5 +2,4 @@ export * from "./rssHandler";
 export * from "./verifyCapacity";
 export * from "./meilisearchClient";
 export * from "./isArchivalTag";
-export * from "./generatePreview";
 export * from "./utils";
