+++ packages/lib/package.json
@@ -4,13 +4,14 @@
   "private": true,
   "main": "index.ts",
   "dependencies": {
+    "@linkwarden/filesystem": "*",
     "@linkwarden/prisma": "*",
     "@linkwarden/types": "*",
-    "@linkwarden/filesystem": "*",
+    "clsx": "^2.1.1",
     "jimp": "^0.22.10",
     "meilisearch": "^0.48.2",
     "rss-parser": "^3.13.0",
-    "clsx": "^2.1.1",
+    "sharp": "^0.34.3",
     "tailwind-merge": "^3.3.0"
   }
 }
